feat: use study session filters in card list
