+++ src/app/i18n/resources.ts
@@ -274,14 +274,18 @@ export const resources = {
         },
         tagFilter: {
           title: "Tags",
-          description: "Choose which tagged cards belong in this session.",
-          all: "All",
+          noFilter: "No filter",
+          selected_one: "{{count}} selected",
+          selected_other: "{{count}} selected",
           clear: "Clear",
-          matchAll: "Match all selected tags",
-          everySelected: "Cards must include every selected tag.",
-          anySelected: "Cards can include any selected tag.",
-          conjunctionAnd: "AND",
-          conjunctionOr: "OR",
+          match: "Match",
+          any: "Any",
+          all: "All",
+          choicesAria: "Tag choices",
+          empty: "No tags available.",
+          showFewer: "Show fewer tags",
+          showMore_one: "Show {{count}} more tag",
+          showMore_other: "Show {{count}} more tags",
         },
       },
       studyStart: {
@@ -302,21 +306,6 @@ export const resources = {
         deckNotFound: {
           title: "Deck not found",
           description: "The requested deck is unavailable or has been removed.",
-        },
-        tags: {
-          title: "Tags",
-          noFilter: "No filter",
-          selected_one: "{{count}} selected",
-          selected_other: "{{count}} selected",
-          clear: "Clear",
-          match: "Match",
-          any: "Any",
-          all: "All",
-          choicesAria: "Tag choices",
-          empty: "No tags available.",
-          showFewer: "Show fewer tags",
-          showMore_one: "Show {{count}} more tag",
-          showMore_other: "Show {{count}} more tags",
         },
       },
       studySession: {
@@ -853,14 +842,18 @@ export const resources = {
         },
         tagFilter: {
           title: "タグ",
-          description: "このセッションに含めるタグ付きカードを選択します。",
-          all: "すべて",
+          noFilter: "フィルターなし",
+          selected_one: "{{count}}個を選択中",
+          selected_other: "{{count}}個を選択中",
           clear: "クリア",
-          matchAll: "選択したすべてのタグに一致",
-          everySelected: "カードには選択したすべてのタグが必要です。",
-          anySelected: "カードには選択したタグのいずれかが必要です。",
-          conjunctionAnd: "AND",
-          conjunctionOr: "OR",
+          match: "一致条件",
+          any: "いずれか",
+          all: "すべて",
+          choicesAria: "タグの選択肢",
+          empty: "利用できるタグはありません。",
+          showFewer: "表示するタグを減らす",
+          showMore_one: "ほか{{count}}個のタグを表示",
+          showMore_other: "ほか{{count}}個のタグを表示",
         },
       },
       studyStart: {
@@ -881,21 +874,6 @@ export const resources = {
         deckNotFound: {
           title: "デッキが見つかりません",
           description: "指定されたデッキは利用できないか、削除されています。",
-        },
-        tags: {
-          title: "タグ",
-          noFilter: "フィルターなし",
-          selected_one: "{{count}}個を選択中",
-          selected_other: "{{count}}個を選択中",
-          clear: "クリア",
-          match: "一致条件",
-          any: "いずれか",
-          all: "すべて",
-          choicesAria: "タグの選択肢",
-          empty: "利用できるタグはありません。",
-          showFewer: "表示するタグを減らす",
-          showMore_one: "ほか{{count}}個のタグを表示",
-          showMore_other: "ほか{{count}}個のタグを表示",
         },
       },
       studySession: {
