+++ src/features/deck-filter/index.ts
@@ -1,3 +1,2 @@
 export { DeckFilterForm } from "./ui/DeckFilterForm";
-export { DifficultyRange } from "./ui/DifficultyRange";
 export { useDeckFilterState } from "./model/useDeckFilterState";
