--- src/pages/study-session-start/ui/StudySessionFilters.stories.tsx
@@ -1,96 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import * as React from "react";
-import { fn } from "storybook/test";
-
-import { StudySessionFilters } from "./StudySessionFilters";
-
-type StudySessionFiltersProps = React.ComponentProps<typeof StudySessionFilters>;
-
-const tags = Array.from({ length: 12 }, (_, index) => `tag-${String(index + 1)}`);
-
-const InteractiveStudySessionFilters: React.FC<StudySessionFiltersProps> = (props) => {
-  const [difficultyMax, setDifficultyMax] = React.useState(props.difficultyMax);
-  const [difficultyMin, setDifficultyMin] = React.useState(props.difficultyMin);
-  const [selectedTags, setSelectedTags] = React.useState(props.selectedTags);
-  const [tagAndFilter, setTagAndFilter] = React.useState(props.tagAndFilter);
-
-  return (
-    <StudySessionFilters
-      {...props}
-      difficultyMax={difficultyMax}
-      difficultyMin={difficultyMin}
-      selectedTags={selectedTags}
-      tagAndFilter={tagAndFilter}
-      clearDifficultyRange={() => {
-        props.clearDifficultyRange();
-        setDifficultyMax(null);
-        setDifficultyMin(null);
-      }}
-      setDifficultyMax={(value) => {
-        props.setDifficultyMax(value);
-        setDifficultyMax(value);
-      }}
-      setDifficultyMin={(value) => {
-        props.setDifficultyMin(value);
-        setDifficultyMin(value);
-      }}
-      setSelectedTags={(value) => {
-        props.setSelectedTags(value);
-        setSelectedTags(value);
-      }}
-      setTagAndFilter={(value) => {
-        props.setTagAndFilter(value);
-        setTagAndFilter(value);
-      }}
-    />
-  );
-};
-
-const meta = {
-  title: "Pages/Study Session Start/StudySessionFilters",
-  component: StudySessionFilters,
-  tags: ["autodocs"],
-  args: {
-    difficultyLowerBound: 1,
-    difficultyMax: 8,
-    difficultyMin: 3,
-    difficultyUpperBound: 10,
-    selectedTags: ["tag-12", "tag-3"],
-    tagAndFilter: false,
-    tags,
-    dirty: true,
-    saving: false,
-    save: fn(async () => undefined),
-    clearDifficultyRange: fn(),
-    setDifficultyMax: fn(),
-    setDifficultyMin: fn(),
-    setSelectedTags: fn(),
-    setTagAndFilter: fn(),
-  },
-  render: (args) => <InteractiveStudySessionFilters {...args} />,
-} satisfies Meta<typeof StudySessionFilters>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {};
-
-export const NoFilters: Story = {
-  args: { difficultyMax: null, difficultyMin: null, selectedTags: [] },
-};
-
-export const ManyTags: Story = {
-  args: {
-    tags: Array.from({ length: 100 }, (_, index) => `tag-${String(index + 1)}`),
-    selectedTags: Array.from({ length: 12 }, (_, index) => `tag-${String(index + 1)}`),
-    tagAndFilter: true,
-  },
-};
-
-export const Mobile320: Story = {
-  globals: { viewport: { value: "iphone5", isRotated: false } },
-};
-
-export const Dark: Story = {
-  globals: { theme: "dark" },
-};
