+++ src/pages/study-session-start/ui/StudySessionStartPage.tsx
@@ -5,13 +5,12 @@ import { useKey } from "react-use";
 
 import type { Deck } from "@/entities/deck";
 import { useDeck } from "@/entities/deck";
-import { useDeckFilterState } from "@/features/deck-filter";
+import { DeckFilterForm, useDeckFilterState } from "@/features/deck-filter";
 import { routes } from "@/shared/router";
 import { AppLayout } from "@/widgets/app-layout";
 import { RouteNotFound } from "@/widgets/route-not-found";
 
 import { useStudySessionStartState } from "../model/useStudySessionStartState";
-import { StudySessionFilters } from "./StudySessionFilters";
 import { StudySessionStart } from "./StudySessionStart";
 
 // The Enter shortcut listens at the window level, so interactive controls must own the event.
@@ -47,7 +46,7 @@ const AvailableStudySessionStartPage: React.FC<{ deck: Deck }> = ({ deck }) => {
         cardsLength={state.cardsLength}
         disabled={filter.saving}
         onClickStart={start}
-        filterSlot={<StudySessionFilters {...filter} tags={state.tags} />}
+        filterSlot={<DeckFilterForm {...filter} tags={state.tags} />}
       />
     </AppLayout>
   );
